backend/requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# OpenAI SDK (1.14.0+ required for word-level timestamps)
openai>=1.14.0

# Replicate SDK for AI video generation
replicate==0.25.1

# Audio Processing
pydub==0.25.1

# Image Processing
Pillow>=10.0.0

# Audio Metadata
mutagen>=1.47.0

# Manim (Mathematical Animation Engine)
manim==0.18.0

# Additional Dependencies
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging and Monitoring
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# System Dependencies Required:
# - ffmpeg (for audio/video processing)
# - Python 3.9+
# - LaTeX (for Manim text rendering)
#
# Installation instructions:
# macOS: brew install ffmpeg mactex
# Ubuntu: sudo apt-get install ffmpeg texlive texlive-latex-extra
# Windows: Download from ffmpeg.org and install MiKTeX
